Modify load-path logic for compilation and grouping imports
Not entirely sure if `(expand load eval)' is exactly what I want, but
it seems to work; the difficulty here is I'd like to always add the
parent directory of the source location to the load path, and that
relies on some macro-expansion.

Separately, use a mix of `reverse' and `cdr' instead of `drop-right'
to avoid importing SRFI-1 before the main module import.